requirements.txt
pytorch==2.0.1
opencv-python==4.7.0.72
pyglm==2.7.0
pyopengl==3.1.0
trimesh==3.8.18